Move more code to utils/analytics.
