+++ Library/Homebrew/cmd/info.rb
@@ -89,7 +89,7 @@ module Homebrew
   def print_info
     if ARGV.named.empty?
       if args.analytics?
-        output_analytics
+        Utils::Analytics.output
       elsif HOMEBREW_CELLAR.exist?
         count = Formula.racks.length
         puts "#{count} #{"keg".pluralize(count)}, #{HOMEBREW_CELLAR.dup.abv}"
@@ -104,13 +104,13 @@ module Homebrew
             Formulary.find_with_priority(f)
           end
           if args.analytics?
-            output_formula_analytics(formula)
+            Utils::Analytics.formula_output(formula)
           else
             info_formula(formula)
           end
         rescue FormulaUnavailableError => e
           if args.analytics?
-            output_analytics(filter: f)
+            Utils::Analytics.output(filter: f)
             next
           end
           ofail e.message
@@ -234,169 +234,7 @@ module Homebrew
     caveats = Caveats.new(f)
     ohai "Caveats", caveats.to_s unless caveats.empty?
 
-    output_formula_analytics(f)
-  end
-
-  def formulae_api_json(endpoint)
-    return if ENV["HOMEBREW_NO_ANALYTICS"] || ENV["HOMEBREW_NO_GITHUB_API"]
-
-    output, = curl_output("--max-time", "5",
-                          "https://formulae.brew.sh/api/#{endpoint}")
-    return if output.blank?
-
-    JSON.parse(output)
-  rescue JSON::ParserError
-    nil
-  end
-
-  def analytics_table(category, days, results, os_version: false, cask_install: false)
-    oh1 "#{category} (#{days} days)"
-    total_count = results.values.inject("+")
-    formatted_total_count = format_count(total_count)
-    formatted_total_percent = format_percent(100)
-
-    index_header = "Index"
-    count_header = "Count"
-    percent_header = "Percent"
-    name_with_options_header = if os_version
-      "macOS Version"
-    elsif cask_install
-      "Token"
-    else
-      "Name (with options)"
-    end
-
-    total_index_footer = "Total"
-    max_index_width = results.length.to_s.length
-    index_width = [
-      index_header.length,
-      total_index_footer.length,
-      max_index_width,
-    ].max
-    count_width = [
-      count_header.length,
-      formatted_total_count.length,
-    ].max
-    percent_width = [
-      percent_header.length,
-      formatted_total_percent.length,
-    ].max
-    name_with_options_width = Tty.width -
-                              index_width -
-                              count_width -
-                              percent_width -
-                              10 # spacing and lines
-
-    formatted_index_header =
-      format "%#{index_width}s", index_header
-    formatted_name_with_options_header =
-      format "%-#{name_with_options_width}s",
-             name_with_options_header[0..name_with_options_width-1]
-    formatted_count_header =
-      format "%#{count_width}s", count_header
-    formatted_percent_header =
-      format "%#{percent_width}s", percent_header
-    puts "#{formatted_index_header} | #{formatted_name_with_options_header} | "\
-         "#{formatted_count_header} |  #{formatted_percent_header}"
-
-    columns_line = "#{"-"*index_width}:|-#{"-"*name_with_options_width}-|-"\
-                   "#{"-"*count_width}:|-#{"-"*percent_width}:"
-    puts columns_line
-
-    index = 0
-    results.each do |name_with_options, count|
-      index += 1
-      formatted_index = format "%0#{max_index_width}d", index
-      formatted_index = format "%-#{index_width}s", formatted_index
-      formatted_name_with_options =
-        format "%-#{name_with_options_width}s",
-               name_with_options[0..name_with_options_width-1]
-      formatted_count = format "%#{count_width}s", format_count(count)
-      formatted_percent = if total_count.zero?
-        format "%#{percent_width}s", format_percent(0)
-      else
-        format "%#{percent_width}s",
-               format_percent((count.to_i * 100) / total_count.to_f)
-      end
-      puts "#{formatted_index} | #{formatted_name_with_options} | " \
-           "#{formatted_count} | #{formatted_percent}%"
-      next if index > 10
-    end
-    return unless results.length > 1
-
-    formatted_total_footer =
-      format "%-#{index_width}s", total_index_footer
-    formatted_blank_footer =
-      format "%-#{name_with_options_width}s", ""
-    formatted_total_count_footer =
-      format "%#{count_width}s", formatted_total_count
-    formatted_total_percent_footer =
-      format "%#{percent_width}s", formatted_total_percent
-    puts "#{formatted_total_footer} | #{formatted_blank_footer} | "\
-         "#{formatted_total_count_footer} | #{formatted_total_percent_footer}%"
-  end
-
-  def output_analytics(filter: nil)
-    days = args.days || "30"
-    category = args.category || "install"
-    json = formulae_api_json("analytics/#{category}/#{days}d.json")
-    return if json.blank? || json["items"].blank?
-
-    os_version = category == "os-version"
-    cask_install = category == "cask-install"
-    results = {}
-    json["items"].each do |item|
-      key = if os_version
-        item["os_version"]
-      elsif cask_install
-        item["cask"]
-      else
-        item["formula"]
-      end
-      if filter.present?
-        next if key != filter && !key.start_with?("#{filter} ")
-      end
-      results[key] = item["count"].tr(",", "").to_i
-    end
-
-    if filter.present? && results.blank?
-      onoe "No results matching `#{filter}` found!"
-      return
-    end
-
-    analytics_table(category, days, results, os_version: os_version, cask_install: cask_install)
-  end
-
-  def output_formula_analytics(f)
-    json = formulae_api_json("#{formula_path}/#{f}.json")
-    return if json.blank? || json["analytics"].blank?
-
-    full_analytics = args.analytics? || args.verbose?
-
-    ohai "Analytics"
-    json["analytics"].each do |category, value|
-      category = category.tr("_", "-")
-      analytics = []
-
-      value.each do |days, results|
-        days = days.to_i
-        if full_analytics
-          if args.days.present?
-            next if args.days&.to_i != days
-          end
-          if args.category.present?
-            next if args.category != category
-          end
-
-          analytics_table(category, days, results)
-        else
-          total_count = results.values.inject("+")
-          analytics << "#{number_readable(total_count)} (#{days} days)"
-        end
-      end
-
-      puts "#{category}: #{analytics.join(", ")}" unless full_analytics
-    end
+    Utils::Analytics.formula_output(f)
   end
 
   def decorate_dependencies(dependencies)
@@ -423,24 +261,4 @@ module Homebrew
 
     "#{dep.name} #{dep.option_tags.map { |o| "--#{o}" }.join(" ")}"
   end
-
-  def format_count(count)
-    count.to_s.reverse.gsub(/(\d{3})(?=\d)/, '\\1,').reverse
-  end
-
-  def format_percent(percent)
-    format("%<percent>.2f", percent: percent)
-  end
-
-  def formula_path
-    "formula"
-  end
-  alias_method :generic_formula_path, :formula_path
-
-  def analytics_path
-    "analytics"
-  end
-  alias_method :generic_analytics_path, :analytics_path
-
-  require "extend/os/cmd/info"
 end
