--- Library/Homebrew/extend/os/cmd/info.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require "extend/os/linux/info" if OS.linux?
