+++ Library/Homebrew/extend/os/utils/analytics.rb
@@ -1,3 +1,4 @@
 # frozen_string_literal: true
 
+require "extend/os/linux/utils/analytics" if OS.linux?
 require "extend/os/mac/utils/analytics" if OS.mac?
