+++ Library/Homebrew/test/cmd/analytics_spec.rb
@@ -14,7 +14,7 @@ describe "brew analytics", :integration_test do
 
     brew "analytics", "off"
     expect { brew "analytics", "HOMEBREW_NO_ANALYTICS" => nil }
-      .to output(/Analytics is disabled/).to_stdout
+      .to output(/Analytics are disabled/).to_stdout
       .and not_to_output.to_stderr
       .and be_a_success
   end
